Changed the facet order in % over time graph
David wanted to talk about the 10K pop size run before talking about
the three 1K runs in the % of successful events over time graph. I
re-ordered the facets, added “PopSize” to the facet label so we know
what the numbers mean, and moved the legend to the bottom of the graph.